Added generate-string-file-with-intermediate option
This option is meant as a solution for the broken Storyboard/xib localization.
You can specify an extra intermediate input whose values are keys to be looked up in the strings input
The output is the intermediate input with the values looked up and translated.

An example of an intermediate input might look like:
"0v1-aK-LCW.normalTitle" = "delete"
"vI7-l7-jAc.normalTitle" = "more_info"

generate-string-file-with-intermediate-option will look up "delete" as a key in the strings input
and replace it with the translated value in the output.